完成了排序框架和排序向量化框架

+++ src/main/scala/com/jetyun/newswire/sorting/SortBean.scala
@@ -7,7 +7,7 @@ import scala.beans.BeanProperty
  * @author 杨勇
  */
 @BeanProperty
-case class HttpPage(id:Long,title:String,content:String,publishTime:Long,keyword:String,url:String){
+case class HttpPage(id:Int,title:String,content:String,publishTime:Long,keyword:String,url:String){
   //定义一个无参的构造函数,方便反射调用
-  def this() = this(1l,"","",1l,"","")
+  def this() = this(1,"","",1l,"","")
 }
